+++ MOD/Systems/MainSystems.cs
@@ -12,47 +12,42 @@ namespace ExtendedRadio.Systems;
 
 public partial class MainSystem : GameSystemBase
 {
-	private bool canLoad = true;
+	//private bool canLoad = true;
 
-    private GameObject gameObject = new();
-    internal static ExtendedRadioMono extendedRadioMono;
+	private readonly GameObject gameObject = new();
+	internal static ExtendedRadioMono extendedRadioMono;
 	internal static NotificationUISystem m_NotificationUISystem;
- 
+
 	protected override void OnCreate()
 	{
 		base.OnCreate();
 		Enabled = false;
-        extendedRadioMono = gameObject.AddComponent<ExtendedRadioMono>();
+		extendedRadioMono = gameObject.AddComponent<ExtendedRadioMono>();
 		m_NotificationUISystem = base.World.GetOrCreateSystemManaged<NotificationUISystem>();
 	}
 
-	protected override void OnUpdate() {}
+	protected override void OnUpdate() { }
 
-	protected override void OnGamePreload(Purpose purpose, GameMode mode)
-	{
-		base.OnGamePreload(purpose, mode);
-		// Print.Info($"OnGamePreload {purpose} | {mode}");
-	}
+	// NO USEFUL ANY MORE.
+	//protected override void OnGameLoadingComplete(Purpose purpose, GameMode mode)
+	//{
+	//	base.OnGameLoadingComplete(purpose, mode);
+	//	// Print.Info($"OnGameLoadingComplete {purpose} | {mode}");
 
-	protected override void OnGameLoadingComplete(Purpose purpose, GameMode mode)
-	{
-		base.OnGameLoadingComplete(purpose, mode);
-		// Print.Info($"OnGameLoadingComplete {purpose} | {mode}");
+	//	if(mode == GameMode.MainMenu) {
+	//		if(canLoad) {
+	//			canLoad = false;
 
-		if(mode == GameMode.MainMenu) {
-			if(canLoad) {
-				canLoad = false;
+	//			//List<string> ModsFolderPaths = [];
+	//			// if(Directory.Exists(EnvPath.kLocalModsPath)) ModsFolderPaths.Add(EnvPath.kLocalModsPath);
+	//			// if(Directory.Exists(Mod.PathToPDXMods)) ModsFolderPaths.Add(Mod.PathToPDXMods);
+	//			//if(Directory.Exists(Mod.ModsFolderCustomRadio)) CustomRadios.RegisterCustomRadioDirectory(Mod.ModsFolderCustomRadio);
 
-				List<string> ModsFolderPaths = [];
-				// if(Directory.Exists(EnvPath.kLocalModsPath)) ModsFolderPaths.Add(EnvPath.kLocalModsPath);
-				// if(Directory.Exists(Mod.PathToPDXMods)) ModsFolderPaths.Add(Mod.PathToPDXMods);
-				if(Directory.Exists(Mod.ModsFolderCustomRadio)) CustomRadios.RegisterCustomRadioDirectory(Mod.ModsFolderCustomRadio);
+	//			//extendedRadioMono.StartCoroutine(CustomRadios.SearchForCustomRadiosFolder(ModsFolderPaths));
 
-				//extendedRadioMono.StartCoroutine(CustomRadios.SearchForCustomRadiosFolder(ModsFolderPaths));
+	//               //Icons.LoadIconsFolder();
 
-                Icons.LoadIconsFolder();
-
-            }
-		}
-	}
+	//           }
+	//	}
+	//}
 }